Changes by itala
