+++ persistencia/src/main/java/br/upe/pojos/Submission.java
@@ -3,6 +3,14 @@ package br.upe.pojos;
 import java.util.UUID;
 
 public class Submission extends EventComponent {
+    private UUID userUuid;
+
+    public void setDescritor(String descritor) {
+        this.descritor = descritor;
+    }
+    public String getDescritor(){
+        return this.descritor;
+    }
     public void setUuid(UUID uuid) {
         this.uuid = uuid;
     }
@@ -15,5 +23,10 @@ public class Submission extends EventComponent {
     public void setEventUuid(UUID eventUuid) {
         this.eventUuid = eventUuid;
     }
-
+    public void setUserUuid(UUID userUuid){
+        this.userUuid = userUuid;
+    }
+    public UUID getUserUuid(){
+        return userUuid;
+    }
 }
